transaction code

+++ QueryProcessor.py
@@ -7,11 +7,12 @@ import DataType
 import Presentation
 
 class QueryProcessor:
-
     def __init__(self):
         self.databaseName = ""
         self.databaseDir = ""
-    
+        self.transaction = False
+
+    #==========================common methods used below===========================================
     def checkIfFilePathExist(self, dirName):
         if os.path.exists(dirName):
             return True
@@ -19,17 +20,17 @@ class QueryProcessor:
             return False
 
     def getDataFromTable(self, tableName):
-        if(self.databaseDir == ""):
-            raise Exception ("Select Database First")
+        if (self.databaseDir == ""):
+            raise Exception("Select Database First")
         tableDir = self.databaseDir + tableName + "/"
         if self.checkIfFilePathExist(tableDir):
             tableDataFilePath = tableDir + tableName + "_data.json"
             tableData = ""
-            with open (tableDataFilePath) as file:
+            with open(tableDataFilePath) as file:
                 tableData = json.load(file)
             return (tableData)
         else:
-            raise Exception (tableName + " Table does not Exist")
+            raise Exception(tableName + " Table does not Exist")
 
     def getValueType(self, data):
         dataType = ["str", "int", "float"]
@@ -40,18 +41,121 @@ class QueryProcessor:
         elif (data == "float"):
             return 1.1
 
+    def getDataInRequiredFormat(self, value):
+        try:
+            data = int(value)
+            return data
+        except ValueError as e:
+            data = float(value)
+            return data
+
+    def saveDataToTable(self, tableName, tableData):
+        tableDataFilePath = self.databaseDir + tableName + "/" + tableName + "_data.json"
+        with open(tableDataFilePath, 'w') as f:
+            f.truncate()
+            json.dump(tableData, f)
+
+    # ============================== commands not afftected by transaction =========================================
+        def useDb(self, DBName):
+            dbDir = "AllDatabase/" + DBName
+            if (self.checkIfFilePathExist(dbDir)):
+                self.databaseName = DBName
+                self.databaseDir = "AllDatabase/" + self.databaseName + "/"
+                print(self.databaseName, " Selected")
+            else:
+                raise Exception("Database " + DBName + " Does Not Exist")
+
+        def createDB(self, DBName):
+            dbDir = "AllDatabase/" + DBName
+            if (self.checkIfFilePathExist(dbDir)):
+                raise Exception("Database already Exist")
+            else:
+                print(DBName, " Created")
+                os.makedirs(dbDir)
+                self.useDb(DBName)
+
+        def dropDb(self, databaseName):
+            databaseDir = "AllDatabase/" + databaseName
+            if (self.checkIfFilePathExist(databaseDir)):
+                shutil.rmtree(databaseDir)
+                print(databaseName + " Database droped")
+            else:
+                raise Exception("Database does not exist " + databaseName)
+
+        def createTable(self, tableName, columnDict, primaryKeyList, foreignKeyList=[]):
+            if (self.databaseName == ""):
+                raise Exception("Select Database first")
+            else:
+                tableName = tableName.rstrip().lstrip()
+                tableDir = self.databaseDir + tableName + "/"
+                tableMetadatFilePath = tableDir + tableName + "_metadata.json"
+                tableDataFilePath = tableDir + tableName + "_data.json"
+
+                if self.checkIfFilePathExist(tableDir):
+                    raise Exception("table already Exist")
+                else:
+                    os.makedirs(tableDir)
+
+                tabelMatadata = dict()
+                tabelMatadata["table_name"] = tableName
+
+                column = dict()
+                for key in columnDict.keys():
+                    columnInputType = DataType.getPythonBasedDataType(columnDict[key])
+                    column[key] = columnInputType
+
+                tabelMatadata["columns"] = column
+                tabelMatadata["primary_key"] = primaryKeyList
+                tabelMatadata["foreign_key"] = foreignKeyList
+                tabelMatadata = json.dumps(tabelMatadata)
+
+                with open(tableMetadatFilePath, 'w') as f:
+                    f.truncate()
+                    f.write(tabelMatadata)
+
+                with open(tableDataFilePath, 'w') as f:
+                    f.truncate()
+                    f.write("[]")
+
+                print(tableName + " Table Created")
+
+        def dropTable(self, tableName):
+            tableDir = self.databaseDir + tableName + "/"
+            if (self.checkIfFilePathExist(tableDir)):
+                shutil.rmtree(tableDir)
+                print(tableName + " table dropped")
+            else:
+                raise Exception("Wrong Table Name")
+
+    #========================================query that dont use Transactions====================================
+    def dropTable(self, tableName):
+        tableDir = self.databaseDir + tableName + "/"
+        if (self.checkIfFilePathExist(tableDir)):
+            shutil.rmtree(tableDir)
+            print(tableName + " table droped")
+        else:
+            raise Exception("Wrong Table Name")
+
     def useDb(self, DBName):
         dbDir = "AllDatabase/" + DBName
-        if( self.checkIfFilePathExist(dbDir)):
+        if (self.checkIfFilePathExist(dbDir)):
             self.databaseName = DBName
             self.databaseDir = "AllDatabase/" + self.databaseName + "/"
             print(self.databaseName, " Selected")
         else:
-            raise Exception("Database " +DBName+" Does Not Exist")
-        
+            raise Exception("Database " + DBName + " Does Not Exist")
+
+    def dropDb(self, databaseName):
+        databaseDir = "AllDatabase/" + databaseName
+        if (self.checkIfFilePathExist(databaseDir)):
+            shutil.rmtree(databaseDir)
+            print(databaseName + " Database droped")
+        else:
+            raise Exception("Database does not exist " + databaseName)
+
     def createDB(self, DBName):
         dbDir = "AllDatabase/" + DBName
-        if( self.checkIfFilePathExist(dbDir)):
+        if (self.checkIfFilePathExist(dbDir)):
             raise Exception("Database already Exist")
         else:
             os.makedirs(dbDir)
@@ -59,7 +163,7 @@ class QueryProcessor:
             self.useDb(DBName)
 
     def createTable(self, tableName, columnDict, primaryKeyList, foreignKeyList=[]):
-        if(self.databaseName == ""):
+        if (self.databaseName == ""):
             raise Exception("Select Database first")
         else:
             tableName = tableName.rstrip().lstrip()
@@ -94,10 +198,53 @@ class QueryProcessor:
                 f.write("[]")
 
             print(tableName + " Table Created")
+    #================================== Transaction COMMANDS ====================================================
+    def startTransaction(self, DbName, tableName):
+        self.transaction = True
+        self.tableName = tableName
+        self.tableDataList = list()
+        self.savePointDict = dict() #{ 'name': index, }
+        self.useDb(DbName)
+        self.tableDataPath =  "AllDatabase/" + DbName + tableName + "_data.json"
+        self.tableDataList.append(self.getDataFromTable(tableName))
+        print("transaction started")
 
+    def setSavePoint(self, name):
+        self.savePointDict[name] = len(self.tableDataList) - 1
+        print("Save point set")
+
+    def rollback(self, savePoint = ""):
+        if savePoint == "" :
+            self.tableDataList = self.tableDataList[:1]
+            print("rolled back to orignal file")
+        elif savePoint in self.savePointDict.keys():
+            print(self.tableDataList[:self.savePointDict[savePoint]])
+            print(len(self.tableDataList))
+            data = self.tableDataList[:self.savePointDict[savePoint]]
+            print(len(data),"dddddddddddddddddddddddd")
+            self.tableDataList = self.tableDataList[:self.savePointDict[savePoint]]
+            print(len(self.tableDataList))
+            print("rollbacked to save point "+ savePoint)
+        else:
+            raise Exception ("No such savepoint present")
+
+    def commit(self):
+        tableData = self.tableDataList[-1]
+        tableName = self.tableName
+        self.saveDataToTable(tableName, tableData)
+        print("Data Saved to db")
+        self.transaction = False
+
+    def isTransaction(self):
+        return self.transaction
+
+    # ============================== commands afftected by transaction =========================================
     def selectQuery(self, tableName, columnListToDisplay=[], condition={}):
-        tableData = self.getDataFromTable(tableName)
-
+        if self.isTransaction():
+            tableData = self.tableDataList[-1]
+            tableName = self.tableName
+        else:
+            tableData = self.getDataFromTable(tableName)
         if (len(tableData) == 0):
             print(tableName + " Table is Empty")
             return
@@ -121,7 +268,6 @@ class QueryProcessor:
                         elif (operator == ">="):
                             if (str(row[colToCheck]) >= (valueToCheck)):
                                 data.append(row)
-
                         elif (operator == "<="):
                             if (str(row[colToCheck]) <= (valueToCheck)):
                                 data.append(row)
@@ -135,7 +281,6 @@ class QueryProcessor:
                 data.insert(0, tableHeader)
                 '''
                 tableData = data
-
             if ("*" not in (columnListToDisplay)):
                 data = list()
                 for row in tableData:
@@ -151,10 +296,15 @@ class QueryProcessor:
                 '''
                 tableData = data
 
+        #Display to user
         Presentation.displayJSONListData(tableData)
 
-    def deleteQuery(self, tableName, condition = []):
-        tableData = self.getDataFromTable(tableName)
+    def deleteQuery(self, tableName, condition=[]):
+        if self.isTransaction():
+            tableData = self.tableDataList[-1]
+            tableName = self.tableName
+        else:
+            tableData = self.getDataFromTable(tableName)
         if (len(tableData) == 0):
             print(tableName + " Table is Empty")
             return
@@ -171,7 +321,7 @@ class QueryProcessor:
                     if (operator == "="):
                         if (str(row[colToCheck]) == (valueToCheck)):
                             data.remove(row)
-                            i = i -1
+                            i = i - 1
                             dataLength = dataLength - 1
                     elif (operator == ">"):
                         if (str(row[colToCheck]) > (valueToCheck)):
@@ -199,49 +349,73 @@ class QueryProcessor:
                             i = i - 1
                             dataLength = dataLength - 1
                 i = i + 1
-
             tableData = data
-        tableDir = self.databaseDir + tableName + "/"
-        tableDataFilePath = tableDir + tableName + "_data.json"
-        with open(tableDataFilePath, 'w') as f:
-            f.truncate()
-            if(len(condition) > 0):
-                json.dump(tableData, f)
-            else:
-                f.write("[]")
 
-    def insertQuery(self, tableName, valueList, colList = []):
-        tableData = self.getDataFromTable(tableName)
+        if self.isTransaction():
+            self.tableDataList.append(tableData)
+        else:
+            tableDir = self.databaseDir + tableName + "/"
+            tableDataFilePath = tableDir + tableName + "_data.json"
+            with open(tableDataFilePath, 'w') as f:
+                f.truncate()
+                if (len(condition) > 0):
+                    json.dump(tableData, f)
+                else:
+                    f.write("[]")
+
+    def insertQuery(self, tableName, valueList, colList=[]):
+        if self.isTransaction():
+            tableData = self.tableDataList[-1]
+            tableName = self.tableName
+        else:
+            tableData = self.getDataFromTable(tableName)
         tableDir = self.databaseDir + tableName + "/"
         tableDataFilePath = tableDir + tableName + "_metadata.json"
         metaData = ""
         with open(tableDataFilePath) as file:
             metaData = json.load(file)
         metaData = metaData["columns"]
-        if( len(colList) == 0):
-            if( len(valueList) == len(metaData)):
+        if (len(colList) == 0):
+            if (len(valueList) == len(metaData)):
                 i = 0
                 dataRow = dict()
                 for key in metaData.keys():
-                    if(type(valueList[i]) == type(self.getValueType(metaData[key]))):
-                        dataRow[key] = valueList[i]
+                    if type(self.getValueType(metaData[key])) != type("str"):
+                        try:
+                            dataInNeededFormat = self.getDataInRequiredFormat(valueList[i])
+                        except:
+                            raise Exception(str(key) + " should be of type " + str(type(self.getValueType(metaData[key]))))
+                        if (type(dataInNeededFormat) == type(self.getValueType(metaData[key]))):
+                            dataRow[key] = dataInNeededFormat
+                        else:
+                            raise Exception(str(key) + " should be of type " + str(type(self.getValueType(metaData[key]))))
                     else:
-                        raise Exception (str(key) + " should be of type " + str(type(self.getValueType(metaData[key]))))
+                        # its in str format so just put it in
+                        dataRow[key] = valueList[i]
                     i = i + 1
 
                 tableData.append(dataRow)
                 print("data added")
             else:
-                raise Exception( tableName + " have "+ str(len(metaData)) + " columns but "+str(len(valueList)) +" values was given" )
+                raise Exception(tableName + " have " + str(len(metaData)) + " columns but " + str(
+                    len(valueList)) + " values was given")
         else:
             i = 0
             dataRow = dict()
             for col in colList:
+                col = col.lstrip().rstrip()
                 if col in metaData.keys():
-                    if (type(valueList[i]) == type(self.getValueType(metaData[col]))):
-                        dataRow[col] = valueList[i]
+                    if type(self.getValueType(metaData[col])) != type("str"):
+                        try:
+                            dataInNeededFormat = self.getDataInRequiredFormat(valueList[i])
+                        except:
+                            raise Exception(str(col) + " should be of type " + str(type(self.getValueType(metaData[col]))))
+                        if (type(dataInNeededFormat) == type(self.getValueType(metaData[col]))):
+                            dataRow[col] = dataInNeededFormat
+                        else:
+                            raise Exception(str(col) + " should be of type " + str(type(self.getValueType(metaData[col]))))
                     else:
-                        raise Exception(str(col) + " should be of type " + str(type(self.getValueType(metaData[col]))))
+                        dataRow[col] = valueList[i]
                 i = i + 1
             for key in metaData:
                 if key in dataRow.keys():
@@ -251,27 +425,34 @@ class QueryProcessor:
             tableData.append(dataRow)
             print("data added")
 
-        tableDir = self.databaseDir + tableName + "/"
-        tableDataFilePath = tableDir + tableName + "_data.json"
-        with open(tableDataFilePath, 'w') as f:
-            f.truncate()
-            json.dump(tableData, f)
+        if self.isTransaction():
+            self.tableDataList.append(tableData)
+        else:
+            tableDir = self.databaseDir + tableName + "/"
+            tableDataFilePath = tableDir + tableName + "_data.json"
+            with open(tableDataFilePath, 'w') as f:
+                f.truncate()
+                json.dump(tableData, f)
 
         return metaData
 
+    #method made just of Update query
     def updateRow(self, tableData, i, colList, metaData):
         for key in colList.keys():
-           if key in tableData[i].keys():
-               if (type(colList[key]) == type(self.getValueType(metaData[key]))):
-                   tableData[i][key] = colList[key]
-               else:
-                   raise Exception(str(key) + " should be of type " + str(type(self.getValueType(metaData[key]))))
-           else:
-               raise Exception (key+ " not present in table")
+            if key in tableData[i].keys():
+                if (type(colList[key]) == type(self.getValueType(metaData[key]))):
+                    tableData[i][key] = colList[key]
+                else:
+                    raise Exception(str(key) + " should be of type " + str(type(self.getValueType(metaData[key]))))
+            else:
+                raise Exception(key + " not present in table")
 
-    def updateQuery(self, tableName, colList, condition ):
-        #" UPDATE Customers SET ContactName='Juan' WHERE Country='Mexico';"
-        tableData = self.getDataFromTable(tableName)
+    def updateQuery(self, tableName, colList, condition):
+        if self.isTransaction():
+            tableData = self.tableDataList[-1]
+            tableName = self.tableName
+        else:
+            tableData = self.getDataFromTable(tableName)
 
         tableDir = self.databaseDir + tableName + "/"
         tableDataFilePath = tableDir + tableName + "_metadata.json"
@@ -293,7 +474,7 @@ class QueryProcessor:
                     operator = condition["operator"]
                     valueToCheck = condition["value"]
                     if (colToCheck in row.keys()):
-                        if operator == "=" :
+                        if operator == "=":
                             if (str(row[colToCheck]) == (valueToCheck)):
                                 self.updateRow(tableData, i, colList, metaData)
                         elif operator == ">":
@@ -312,19 +493,13 @@ class QueryProcessor:
                             if (str(row[colToCheck]) != (valueToCheck)):
                                 self.updateRow(tableData, i, colList, metaData)
                     i = i + 1
-
-        tableDir = self.databaseDir + tableName + "/"
-        tableDataFilePath = tableDir + tableName + "_data.json"
-        with open(tableDataFilePath, 'w') as f:
-            f.truncate()
-            json.dump(tableData, f)
+        if self.isTransaction():
+            self.tableDataList.append(tableData)
+        else:
+            tableDir = self.databaseDir + tableName + "/"
+            tableDataFilePath = tableDir + tableName + "_data.json"
+            with open(tableDataFilePath, 'w') as f:
+                f.truncate()
+                json.dump(tableData, f)
 
         print("Table Updated")
-
-    def dropTable(self, tableName):
-        tableDir = self.databaseDir + tableName + "/"
-        if(self.checkIfFilePathExist(tableDir)):
-            shutil.rmtree(tableDir)
-            print(tableName + " table droped")
-        else:
-            raise Exception ("Wrong Table Name")
